+++ algo/trading/executor_exit_handler.py
@@ -531,7 +531,7 @@ class ExitHandler:
             SELECT COALESCE(SUM((details->>'pnl_dollars')::numeric), 0)
             FROM algo_audit_log
             WHERE action_type LIKE 'exit_%%'
-              AND (details->>'trade_id')::bigint = %s
+              AND (details->>'trade_id') = %s
               AND (details->>'full_exit')::boolean = false
             """,
             (trade_id,),
